caizen/gcp_cai_func/requirements.txt
a2wsgi==1.10.7 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.11" and python_version < "4"
build==1.2.2 ; python_version >= "3.11" and python_version < "4.0"
cachecontrol[filecache]==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux") and (sys_platform == "darwin" or platform_python_implementation != "PyPy")
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4"
cloudevents==1.11.0 ; python_version >= "3.11" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4" and platform_system == "Windows" or python_version >= "3.11" and python_version < "4" and os_name == "nt"
crashtest==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
cryptography==43.0.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
deprecated==1.2.14 ; python_version >= "3.11" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.11" and python_version < "4"
distlib==0.3.8 ; python_version >= "3.11" and python_version < "4.0"
dulwich==0.21.7 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.114.1 ; python_version >= "3.11" and python_version < "4.0"
fastjsonschema==2.20.0 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.16.0 ; python_version >= "3.11" and python_version < "4.0"
flask-pydantic==0.12.0 ; python_version >= "3.11" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.11" and python_version < "4"
functions-framework==3.8.1 ; python_version >= "3.11" and python_version < "4"
google-api-core==2.19.2 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-appengine-logging==1.4.5 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-audit-log==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-logging==3.11.2 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-storage==2.18.2 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.66.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "4" and platform_system != "Windows"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.8 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==8.4.0 ; python_version >= "3.11" and python_version < "4.0"
installer==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4"
jaraco-classes==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
jinja2==3.1.4 ; python_version >= "3.11" and python_version < "4"
keyring==24.3.1 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4"
more-itertools==10.5.0 ; python_version >= "3.11" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-api==1.27.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.11" and python_version < "4"
pexpect==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
pkginfo==1.11.1 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.3.2 ; python_version >= "3.11" and python_version < "4.0"
poetry-core==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
poetry-plugin-export==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
poetry==1.8.3 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.28.1 ; python_version >= "3.11" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux") and (sys_platform == "darwin" or platform_python_implementation != "PyPy")
pydantic-core==2.23.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.9.1 ; python_version >= "3.11" and python_version < "4.0"
pyproject-hooks==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
pywin32-ctypes==0.2.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
rapidfuzz==3.9.7 ; python_version >= "3.11" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
secretstorage==3.3.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.38.5 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.13.2 ; python_version >= "3.11" and python_version < "4.0"
trove-classifiers==2024.9.12 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.30.6 ; python_version >= "3.11" and python_version < "4.0"
virtualenv==20.26.4 ; python_version >= "3.11" and python_version < "4.0"
watchdog==5.0.2 ; python_version >= "3.11" and python_version < "4"
werkzeug==3.0.4 ; python_version >= "3.11" and python_version < "4"
wrapt==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
xattr==1.1.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "darwin"
zipp==3.20.1 ; python_version >= "3.11" and python_version < "4.0"
